+++ lib/main.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
-// import 'package:places/mock.dart';
 import 'package:places/ui/res/constants.dart';
-// import 'package:places/ui/screen/sight_details.dart';
 import 'package:places/ui/screen/sight_list_screen.dart';
-//import 'package:places/ui/screen/sight_list_screen.dart';
 
 void main() {
   runApp(const App());
