+++ lib/ui/res/app_assets.dart
@@ -1,0 +1,2 @@
+class AppAssets {
+}
