+++ lib/ui/res/app_colors.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const backgroundColor = Colors.white;
+  static const textColorRegular = Color(0xff3B3E5B);
+  static const textColorLight = Color(0xff7C7E92);
+  static const backgroundPlaceItemBottom = Color(0xfff5f5f5);
+}
