+++ lib/ui/res/app_strings.dart
@@ -1,0 +1,6 @@
+/// Строковые константы приложения
+class AppStrings {
+  static const appTitle = 'Список\nинтересных мест';
+  static const bulidRoute = 'ПОСТРОИТЬ МАРШРУТ';
+  static const closedUntil = 'Закрыто до';
+}
